+++ src/main/java/com/jarviswuod/improvedbillingsystem/dashboard/CustomersDto.java
@@ -1,7 +1,9 @@
 package com.jarviswuod.improvedbillingsystem.dashboard;
 
+import java.math.BigDecimal;
+
 public record CustomersDto(
         String customerName,
-        double totalPaid
+        BigDecimal totalPaid
 ) {
 }
