+++ src/main/java/com/jarviswuod/improvedbillingsystem/dashboard/MonthlyRevenueDto.java
@@ -1,7 +1,9 @@
 package com.jarviswuod.improvedbillingsystem.dashboard;
 
+import java.math.BigDecimal;
+
 public record MonthlyRevenueDto(
         String month,
-        double total
+        BigDecimal total
 ) {
 }
